refactor(cea): rewire headless to use @ai-sdk-tool/headless
